Move AuthCliennt class from api file to separate Auth class and import it from API
