+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/data/api.js
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 "use strict";
-import SwaggerClient from 'swagger-client';
+import SwaggerClient from 'swagger-client'
+import AuthClient from './Auth'
 /**
  * Manage API access keys with corresponding keys,Not related to the keymanager used in backend
  */
@@ -89,96 +90,6 @@ class KeyManager {
     }
 }
 
-class AuthClient {
-
-    static refreshTokenOnExpire() {
-        var timestampSkew = 100;
-        var currentTimestamp = Math.floor(Date.now() / 1000);
-        var tokenTimestamp = localStorage.getItem("expiresIn");
-        var rememberMe = (localStorage.getItem("rememberMe") == 'true');
-        if (rememberMe && (tokenTimestamp - currentTimestamp < timestampSkew)) {
-            var bearerToken = "Bearer " + AuthClient.getCookie("WSO2_AM_REFRESH_TOKEN_1");
-            var loginPromise = authManager.refresh(bearerToken);
-            loginPromise.then(function (data, status, xhr) {
-                authManager.setAuthStatus(true);
-                var expiresIn = data.validityPeriod + Math.floor(Date.now() / 1000);
-                window.localStorage.setItem("expiresIn", expiresIn);
-            });
-            loginPromise.error(
-                function (error) {
-                    var error_data = JSON.parse(error.responseText);
-                    var message = "Error while refreshing token" + "<br/> You will be redirect to the login page ...";
-                    noty({
-                        text: message,
-                        type: 'error',
-                        dismissQueue: true,
-                        modal: true,
-                        progressBar: true,
-                        timeout: 5000,
-                        layout: 'top',
-                        theme: 'relax',
-                        maxVisible: 10,
-                        callback: {
-                            afterClose: function () {
-                                window.location = loginPageUri;
-                            },
-                        }
-                    });
-
-                }
-            );
-        }
-    }
-
-    /**
-     * Static method to handle unauthorized user action error catch, It will look for response status code and skip !401 errors
-     * @param {object} error_response
-     */
-    static unauthorizedErrorHandler(error_response) {
-        if (error_response.status !== 401) { /* Skip unrelated response code to handle in unauthorizedErrorHandler*/
-            throw error_response;
-            /* re throwing the error since we don't handle it here and propagate to downstream error handlers in catch chain*/
-        }
-        let message = "The session has expired" + ".<br/> You will be redirect to the login page ...";
-        if (typeof noty !== 'undefined') {
-            noty({
-                text: message,
-                type: 'error',
-                dismissQueue: true,
-                modal: true,
-                progressBar: true,
-                timeout: 5000,
-                layout: 'top',
-                theme: 'relax',
-                maxVisible: 10,
-                callback: {
-                    afterClose: function () {
-                        window.location = loginPageUri;
-                    },
-                }
-            });
-        } else {
-            throw error_response;
-        }
-    }
-
-    static getCookie(name) {
-        let pairs = document.cookie.split(";");
-        const cookies = {};
-        for (let pair of pairs) {
-            pair = pair.split("=");
-            let cookie_name = pair[0].trim();
-            let value = encodeURIComponent(pair[1]);
-            if (cookie_name === name) {
-                return value;
-            } else {
-                cookies[cookie_name] = value;
-            }
-        }
-        return cookies;
-    }
-
-}
 /**
  * An abstract representation of an API
  */
@@ -197,7 +108,6 @@ class API {
         this.auth_client = new AuthClient();
         this.client.then(
             (swagger) => {
-                // debugger;
                 swagger.setHost("localhost:9292");
                 swagger.setSchemes(["https"]);
                 this.keyMan = new KeyManager(access_key);
